# UniqueApi/Controllers/CitasController.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc;
using Microsoft.EntityFrameworkCore;
using UniqueApi.DAL;
using UniqueApi.Models;

namespace UniqueApi.Controllers
{
    [Route("api/[controller]")]
    [ApiController]
    public class CitasController : ControllerBase
    {
        private readonly Contexto _context;

        public CitasController(Contexto context)
        {
            _context = context;
        }

        // GET: api/Citas
        [HttpGet("/lista/{id}")]
        public async Task<ActionResult<IEnumerable<Cita>>> GetCitas(int id)
        {
          if (_context.Cita == null)
          {
              return NotFound();
          }
            return await _context.Cita.Where(c=> c.ClienteId == id).ToListAsync();
        }

        // GET: api/Citas/5
        [HttpGet("{id}")]
        public async Task<ActionResult<Cita>> GetCita(int id)
        {
          if (_context.Cita == null)
          {
              return NotFound();
          }
            var cita = await _context.Cita.FindAsync(id);

            if (cita == null)
            {
                return NotFound();
            }

            return cita;
        }
        // GET: api/Citas/Top3CitasByClienteId/5
        [HttpGet("Top3CitasByClienteId/{ClienteId}")]
        public async Task<ActionResult<List<Cita>>> Top3CitasByClienteId(int ClienteId)
        {
            if (_context.Cita == null)
            {
                return NotFound();
            }
            var cita = await _context.Cita.Where(c => c.ClienteId == ClienteId && c.EstadoId == 1)
                .OrderBy(t => t.Fecha).Take(3).ToListAsync();

            if (cita == null)
            {
                return NotFound();
            }

            return cita;
        }

        // PUT: api/Citas/5
        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
        [HttpPut("{id}")]
        public async Task<IActionResult> PutCita(int id, Cita cita)
        {
            if (id != cita.CitaId)
            {
                return BadRequest();
            }

            _context.Entry(cita).State = EntityState.Modified;

            try
            {
                await _context.SaveChangesAsync();
            }
            catch (DbUpdateConcurrencyException)
            {
                if (!CitaExists(id))
                {
                    return NotFound();
                }
                else
                {
                    throw;
                }
            }

            return NoContent();
        }

        // POST: api/Citas
        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
        [HttpPost]
        public async Task<ActionResult<Cita>> PostCita(Cita cita)
        {
          if (_context.Cita == null)
          {
              return Problem("Entity set 'Contexto.Cita'  is null.");
          }
            _context.Cita.Add(cita);
            await _context.SaveChangesAsync();

            return CreatedAtAction("GetCita", new { id = cita.CitaId }, cita);
        }

        // DELETE: api/Citas/5
        [HttpDelete("{id}")]
        public async Task<IActionResult> DeleteCita(int id)
        {
            if (_context.Cita == null)
            {
                return NotFound();
            }
            var cita = await _context.Cita.FindAsync(id);
            if (cita == null)
            {
                return NotFound();
            }

            _context.Cita.Remove(cita);
            await _context.SaveChangesAsync();

            return NoContent();
        }

        private bool CitaExists(int id)
        {
            return (_context.Cita?.Any(e => e.CitaId == id)).GetValueOrDefault();
        }
    }
}